+++ ChoGGi's Library/Changelog.md
@@ -1,6 +1,7 @@
 ## Library Changelog
 
 ## v11.4 (Unreleased)
+- Fixed Emergency Lua GC (from this mod at least, thanks Tremualin).
 - FindNearestResource counts resources stored in building depots.
 - Prints a warning if paradox account isn't signed in (mods won't update).